Add capability to capture prefetchdataset to var-opts Refactor code - remove dups
