+++ ngraph_bridge/enable_variable_ops/ngraph_capture_variables.cc
@@ -20,6 +20,7 @@
 #include "ngraph_bridge/enable_variable_ops/ngraph_replace_op_utilities.h"
 #include "ngraph_bridge/ngraph_api.h"
 #include "ngraph_bridge/ngraph_capture_variables.h"
+#include "ngraph_bridge/ngraph_find_replace_prefetchdataset.h"
 #include "ngraph_bridge/ngraph_utils.h"
 
 using namespace std;
@@ -61,6 +62,7 @@ Status CaptureVariables(Graph* graph, std::set<string> skip_these_nodes) {
           {"VariableV2", std::make_pair("NGraphVariable", ReplaceVariable)}};
 
   std::set<Node*> nodes_to_capture;
+  Node* prefetch_node = nullptr;
 
   for (auto node : graph->op_nodes()) {
     std::set<Node*> ref_list;
@@ -84,6 +86,15 @@ Status CaptureVariables(Graph* graph, std::set<string> skip_these_nodes) {
           }
           ref_list.clear();
         }
+      } else if (node->type_string() == "MakeIterator") {
+        if (std::getenv(NGraphPrefetchSharedResouce::NGRAPH_TF_USE_PREFETCH) !=
+            nullptr) {
+          // We expect the MakeIterator to have 1 input thats
+          // an iterator and the other one can be either a
+          // PrefetchDataset node or a ModelDataset node
+          // Other cases are not handled at the moment.
+          prefetch_node = FindPrefetch(node);
+        }
       }
     }
   }
@@ -104,6 +115,18 @@ Status CaptureVariables(Graph* graph, std::set<string> skip_these_nodes) {
   for (auto node : nodes_to_capture) {
     NGRAPH_VLOG(4) << "Removing: " << node->name();
     graph->RemoveNode(node);
+  }
+
+  if (std::getenv(NGraphPrefetchSharedResouce::NGRAPH_TF_USE_PREFETCH) !=
+      nullptr) {
+    if (prefetch_node != nullptr) {
+      return ReplacePrefetch(graph, prefetch_node);
+    } else {
+      return errors::Internal(
+          "Did not find PrefetchDataset or "
+          "ModelDataset+OptimizeDataset+PrefetchDataset as MakeIterator "
+          "nodes' inputs. Only those 2 cases are handled for now.");
+    }
   }
   return Status::OK();
 }
