+++ ngraph_bridge/ngraph_rewrite_pass.cc
@@ -49,41 +49,6 @@ class NGraphRewritePass : public GraphOptimizationPass {
   virtual Status Run(const GraphOptimizationPassOptions& options) = 0;
 
  protected:
-  void DumpGraphs(const GraphOptimizationPassOptions& options, int idx,
-                  std::string filename_prefix, std::string title) {
-    // If we have a "main" graph, dump that.
-    if (options.graph != nullptr) {
-      auto dot_filename = DotFilename(filename_prefix, idx);
-      auto pbtxt_filename = PbtxtFilename(filename_prefix, idx);
-      NGRAPH_VLOG(0) << "Dumping main graph to " << dot_filename;
-      NGRAPH_VLOG(0) << "Dumping main graph to " << pbtxt_filename;
-
-      GraphToDotFile(options.graph->get(), dot_filename, title);
-      GraphToPbTextFile(options.graph->get(), pbtxt_filename);
-    }
-
-    // If we have partition graphs (we shouldn't), dump those.
-    if (options.partition_graphs != nullptr) {
-      int sub_idx = 0;
-
-      for (auto& kv : *options.partition_graphs) {
-        auto dot_filename = DotFilename(filename_prefix, idx, sub_idx);
-        auto pbtxt_filename = PbtxtFilename(filename_prefix, idx, sub_idx);
-        NGRAPH_VLOG(0) << "Dumping subgraph " << sub_idx << " to "
-                       << dot_filename;
-        NGRAPH_VLOG(0) << "Dumping subgraph " << sub_idx << " to "
-                       << pbtxt_filename;
-
-        Graph* pg = kv.second.get();
-
-        GraphToDotFile(pg, dot_filename, title);
-        GraphToPbTextFile(pg, pbtxt_filename);
-
-        sub_idx++;
-      }
-    }
-  }
-
   // Returns a fresh "serial number" to avoid filename collisions in the graph
   // dumps.
   static int FreshIndex() {
